Wrote the SQL query to update a sample
The query updates the sample by adding the lab_technician_id, strain_id and updates the status_id from 1 (Hospital) to 2 (Laboratory) where the sample_id is equal to the one that is entered in the form. It will also display a message that the values have been inserted successfully for the given sample.